Move codeowners checks to the dev folder
